+++ Source/Plugins/Plugin_VideoOGL/Src/GLUtil.cpp
@@ -664,8 +664,8 @@ void OpenGL_Update()
     {
         MValueX = 1.0f / Max;
         MValueY = 1.0f / Max;
-        nXoff = (nBackbufferWidth - (640 * MValueX)) / 2;
-        nYoff = (nBackbufferHeight - (480 * MValueY)) / 2;
+        nXoff = (int)((nBackbufferWidth - (640 * MValueX)) / 2);
+        nYoff = (int)((nBackbufferHeight - (480 * MValueY)) / 2);
     }
 	
 	// tell the debugger
